New XQuery 1.1 Functions added (II):
- fn:format-integer($value as xs:integer?, $picture as xs:string) as xs:string
- fn:format-integer($value as xs:integer?, $picture as xs:string,
    $language as xs:language) as xs:string

Regular expression functions:
- -q: literal option

XQuery:
- converted some int to long values

+++ src/main/java/org/basex/api/dom/BXNList.java
@@ -48,6 +48,6 @@ public class BXNList implements NodeList {
 
   @Override
   public int getLength() {
-    return xquery != null ? xquery.size() : nodes.size();
+    return (int) (xquery != null ? xquery.size() : nodes.size());
   }
 }
