+++ src/main/java/org/basex/query/item/Item.java
@@ -54,7 +54,7 @@ public abstract class Item extends Expr {
       @Override
       public Item next() { return (more ^= true) ? Item.this : null; }
       @Override
-      public int size() { return 1; }
+      public long size() { return 1; }
       @Override
       public Item get(final long i) { return i == 0 ? Item.this : null; }
     };
